LPS-69239 Move to util package
